tried to make some changes for SPhase

+++ main/UseCases/FinalReportMaker.java
@@ -30,7 +30,7 @@ public class FinalReportMaker implements ReportMaker {
         String internNames = currentHRSystem.getInternNames();
 
         ArrayList<Project> projList = currentHRSystem.getProject(currentMonth);
-        HashMap<String, Integer> projectCompatibilityList = new HashMap<>();
+        HashMap<String, Float> projectCompatibilityList = new HashMap<>();
         for (Project proj : projList) {
             projectCompatibilityList.putAll(proj.getSkillsCompatibilities());
         }
@@ -65,7 +65,7 @@ public class FinalReportMaker implements ReportMaker {
     }
 
     @Override
-    public String bakeInternsPerformances (String internNames, ArrayList<HashMap<String, Double>>  internSkills, HashMap<String, Integer> projectSkill) {
+    public String bakeInternsPerformances (String internNames, ArrayList<HashMap<String, Double>>  internSkills, HashMap<String, Float> projectSkill) {
         StringBuilder returnLine = new StringBuilder(GamePrompts.INTERN_PERFORMANCE_HEADER + internNames + "\n");
         String[] internNamesList = internNames.split("|");
         for (int i = 0; i != internNamesList.length; i+=1) {
@@ -75,7 +75,7 @@ public class FinalReportMaker implements ReportMaker {
     }
 
     @Override
-    public int calculateInternPerformance(HashMap<String, Double> internSkills, HashMap<String, Integer> projectSkill) {
+    public int calculateInternPerformance(HashMap<String, Double> internSkills, HashMap<String, Float> projectSkill) {
         int result = 0;
         ArrayList<Double> effectiveSkills = new ArrayList<Double>();
         for (String key : internSkills.keySet()) {
